Code review fixes!
The biggest functionality change here is we no longer support streaming
uploads. We simply load the file into memory, sign the request and send
it in the body of the request.

Fix grammatical, JSDoc, and style guide errors.
Suffixed functions that return a promise with Async.
Added a watch script.